+++ students/21K654/21K654-p02/src/main/java/ru/mirea/practice/Task3/Circle.java
@@ -1,0 +1,25 @@
+package ru.mirea.practice.Task3;
+
+public class Circle {
+    private Point center;
+    private double radius;
+    public Circle(double x, double y){
+        this.center = new Point(x,y);
+    }
+    public Circle(double x, double y,double r){
+        this.center = new Point(x,y);
+        this.radius = r;
+    }
+    public Circle(){
+        this.center = new Point();
+    }
+
+    public void setCenter(Point center) {
+        this.center = center;
+    }
+
+    @Override
+    public String toString(){
+        return center.toString() + " " + this.radius;
+    }
+}
