+++ students/21K654/21K654-p02/src/main/java/ru/mirea/practice/Task3/Point.java
@@ -1,0 +1,39 @@
+package ru.mirea.practice.Task3;
+
+public class Point {
+    private double x;
+    private  double y;
+    public Point(double x, double y){
+        this.x = x; this.y = y;
+    }
+    public Point(){
+        this.x = 0;this.y = 0;
+    }
+
+    public double getX() {
+        return x;
+    }
+
+    public double getY() {
+        return y;
+    }
+
+    public void setX(double x) {
+        this.x = x;
+    }
+
+    public void setY(double y) {
+        this.y = y;
+    }
+    public void setXY(double x, double y){
+        this.x = x; this.y = y;
+    }
+    public void move(double xDisp, double yDisp){
+        this.x += xDisp;this.y += yDisp;
+    }
+    @Override
+    public  String toString(){
+        return this.x + " " + this.y;
+    }
+
+}
