+++ students/21K654/21K654-p02/src/main/java/ru/mirea/practice/Task4/Input.java
@@ -1,0 +1,7 @@
+package ru.mirea.practice.Task4;
+
+import java.util.Scanner;
+
+public interface Input {
+    void input(Scanner sc);
+}
