+++ students/21K654/21K654-p02/src/main/java/ru/mirea/practice/Task7/Book.java
@@ -1,0 +1,13 @@
+package ru.mirea.practice.Task7;
+
+import ru.mirea.practice.Task1.Author;
+
+public class Book {
+    private Author author;
+    private int year;
+    private String title;
+
+    public int getYear() {
+        return year;
+    }
+}
