+++ students/21K654/21K654-p02/src/main/java/ru/mirea/practice/Task7/BookTest.java
@@ -1,0 +1,7 @@
+package ru.mirea.practice.Task7;
+
+public class BookTest {
+    public static void main(String[] args) {
+
+    }
+}
