--- students/21K654/21K654-p02/src/main/java/ru/mirea/practice/Task_1/TestAuthor.java
@@ -1,10 +1,0 @@
-package ru.mirea.practice.Task_1;
-
-public class TestAuthor {
-    public static void main(String[] args) {
-        Author author = new Author("zhigga","gog@mail.ru",'m');
-        System.out.println(author.getEmail() + "   " + author.getName() + "   " + author.getGender());
-        author.setEmail("gig@mail.ru");
-        System.out.println(author.getEmail() + '\n' + author.toString());
-    }
-}
