--- students/21K654/21K654-p02/src/main/java/ru/mirea/practice/Task_6/Circle.java
@@ -1,24 +1,0 @@
-package ru.mirea.practice.Task_6;
-
-public class Circle {
-    private double radius;
-    public Circle(double radius){
-        this.radius = radius;
-    }
-    public double square(){
-        return Math.PI * Math.pow(radius,2.0);
-    }
-    public double length(){
-        return 2 * radius * Math.PI;
-    }
-
-    public void setRadius(double radius) {
-        this.radius = radius;
-    }
-    public int toCompare(Circle a){
-        return (int)(this.radius - a.getRadius());
-    }
-    public double getRadius() {
-        return radius;
-    }
-}
