--- students/21K654/21K654-p02/src/main/java/ru/mirea/practice/Task_7/BookTest.java
@@ -1,7 +1,0 @@
-package ru.mirea.practice.Task_7;
-
-public class BookTest {
-    public static void main(String[] args) {
-
-    }
-}
