+++ students/21K654/21K654-p03/src/main/java/ru/mirea/practice/Task2/Circle1.java
@@ -1,0 +1,4 @@
+package ru.mirea.practice.Task2;
+
+public class Circle1 {
+}
