--- students/21K654/21K654-p03/src/main/java/ru/mirea/practice/Task_2/Circle1.java
@@ -1,4 +1,0 @@
-package ru.mirea.practice.Task_2;
-
-public class Circle1 {
-}
